+++ NEWS.md
@@ -1,3 +1,12 @@
+This file lists changes in the R package for each release.
+
+# ShrinkNet 1.0.2
+
+* improved code syntax
+* changed \code{.convertToTime()} to \code{.convert2time()}
+* improved documentation
+* \code{ShrinkNet()} function argument checking using \code{assertthat}
+
 # ShrinkNet 1.0.1
 
 * package updated to pass R CMD check